logic improvement update IPKSM.py
removed change to folder in 5.2 chunk check
put the chunk check at the start as step 0 so that ndwi doesn't have to be calculated if they're all already there
now asking the user if they want to continue regardless of the percentage completion of the chunks (unless completely finished). if chunk generation is complete, none of the other steps get carried out (because what's the point) and they are all placed under an if statement.