Update IPKSM.py
finished section "6. Load Model"
added assignment of "class_names" list but this will be amended with an additional "neither" class tomorrow morning
the ndwi model epochs-30_20250421_230014 model was trained using the fixed 0.4 max multiplier in ipdgs. slightly better results
i'm going to need to add the third class back. i'll just make it more reasonably sized. using this, and ipdgs and ipksm_trainer as they are now, i've got a 30 epoch model that is heading in the right direction. next step is the third class and i'm going to try to make the margins much much bigger for training data.